perf(stake): cache + multicall the orbit graph; fix stale dialog copy
The orbit did dozens of sequential client RPC round-trips on every mount
(per-backer balanceOf + convertToAssets in a loop, per-pool log scans),
so it painted slowly. Move it server-side and cache it:

- New /api/stake-graph (revalidate 60s + stale-while-revalidate 300s), backed
  by src/services/stake-graph.ts. Shared across users → warm loads are instant.
- Multicall3: each vault's totalAssets/totalSupply/every balance goes in ONE
  aggregated call; shares→assets is computed locally, dropping all per-backer
  convertToAssets calls. MOR log scans + usersData reads are parallelized and
  multicalled too.
- useStakeGraph is now a thin react-query hook over the cached route (kept warm
  between navigations); same public API and types.

Also fixes wrong dialog copy: the disclaimer no longer claims staking "isn't
wired to a contract yet" (it is), and the intro no longer says "Stake ETH".
Localized en + pt-br.

Multicall verified against a live vault; tsc + next build clean.
