+++ messages/en/stake.json
@@ -7,7 +7,7 @@
   "selectedRider": "{name} is ready to drop in.",
   "stakeCta": "Stake with {name}",
   "stakeToast": "Staked with {name}",
-  "dialogIntro": "Stake ETH as {name}. Your ETH keeps earning staking yield — the rewards stream to you, your rider, and the Gnars treasury.",
+  "dialogIntro": "Back {name}. Your deposit stays yours and keeps earning — only the yield is shared with your rider and the Gnars treasury.",
   "amountLabel": "Amount to stake",
   "aprNote": "at {rate}% {rateType} via {source}",
   "flowTitle": "How rewards flow",
@@ -18,7 +18,7 @@
   "treasuryLabel": "Gnars Treasury",
   "confirm": "Confirm stake",
   "cancel": "Cancel",
-  "disclaimer": "Preview only — staking isn't wired to a contract yet, and this isn't financial advice.",
+  "disclaimer": "Your deposit stays yours and is withdrawable anytime. Rates vary; this isn't financial advice.",
   "prev": "Previous rider",
   "next": "Next rider",
   "statusTitle": "Status",
