+++ messages/pt-br/stake.json
@@ -7,7 +7,7 @@
   "selectedRider": "{name} está pronto para descer.",
   "stakeCta": "Fazer stake com {name}",
   "stakeToast": "Stake feito com {name}",
-  "dialogIntro": "Faça stake de ETH como {name}. Seu ETH continua rendendo staking — as recompensas fluem para você, seu rider e o tesouro da Gnars.",
+  "dialogIntro": "Apoie {name}. Seu depósito continua seu e rendendo — só o rendimento é dividido com o rider e o tesouro da Gnars.",
   "amountLabel": "Valor para stake",
   "aprNote": "a {rate}% {rateType} via {source}",
   "flowTitle": "Como as recompensas fluem",
@@ -18,7 +18,7 @@
   "treasuryLabel": "Tesouro Gnars",
   "confirm": "Confirmar stake",
   "cancel": "Cancelar",
-  "disclaimer": "Apenas prévia — o stake ainda não está ligado a um contrato, e isto não é aconselhamento financeiro.",
+  "disclaimer": "Seu depósito continua seu e pode ser sacado quando quiser. As taxas variam; isto não é aconselhamento financeiro.",
   "prev": "Rider anterior",
   "next": "Próximo rider",
   "statusTitle": "Status",
